+++ packages/components/board/examples/custom-template/custom-template.component.ts
@@ -1,0 +1,24 @@
+import { Component, OnInit } from '@angular/core';
+import { ThyBoardCard, ThyBoardEntry, ThyBoardLane } from '@tethys/pro/board';
+import { entries, items, lanes } from '../mock';
+
+interface CardInfo extends ThyBoardCard {
+    title: string;
+}
+
+@Component({
+    selector: 'thy-pro-board-custom-template-example',
+    templateUrl: './custom-template.component.html',
+    styleUrls: ['./custom-template.component.scss']
+})
+export class ThyProBoardCustomTemplateExampleComponent implements OnInit {
+    entries: ThyBoardEntry[] = [...entries];
+
+    items: CardInfo[] = [...(items as CardInfo[])];
+
+    lanes: ThyBoardLane[] = [...lanes];
+
+    constructor() {}
+
+    ngOnInit(): void {}
+}