+++ packages/components/board/examples/mock.ts
@@ -1,0 +1,152 @@
+import { ThyBoardCard, ThyBoardEntry, ThyBoardLane } from '@tethys/pro/board';
+
+export const entries: ThyBoardEntry[] = [
+    {
+        _id: '1',
+        name: '未开始'
+    },
+    {
+        _id: '2',
+        name: '进行中'
+    },
+    {
+        _id: '3',
+        name: '已完成'
+    },
+    {
+        _id: '4',
+        name: '已关闭'
+    }
+];
+
+export const items: ThyBoardCard[] = [
+    {
+        _id: '1',
+        title: '项目1',
+        laneId: '1',
+        entryId: '1'
+    },
+    {
+        _id: '2',
+        title: '项目2',
+        laneId: '1',
+        entryId: '1'
+    },
+    {
+        _id: '3',
+        title: '项目3',
+        laneId: '2',
+        entryId: '2'
+    },
+    {
+        _id: '4',
+        title: '项目4',
+        laneId: '1',
+        entryId: '1'
+    },
+    {
+        _id: '5',
+        title: '项目5',
+        laneId: '1',
+        entryId: '1'
+    },
+    {
+        _id: '6',
+        title: '项目6',
+        laneId: '1',
+        entryId: '1'
+    },
+    {
+        _id: '7',
+        title: '项目7',
+        laneId: '1',
+        entryId: '1'
+    },
+    {
+        _id: '8',
+        title: '项目8',
+        laneId: '1',
+        entryId: '1'
+    },
+    {
+        _id: '9',
+        title: '项目9',
+        laneId: '2',
+        entryId: '2'
+    },
+    {
+        _id: '10',
+        title: '项目10',
+        laneId: '1',
+        entryId: '1'
+    },
+    {
+        _id: '11',
+        title: '项目11',
+        laneId: '1',
+        entryId: '1'
+    },
+    {
+        _id: '12',
+        title: '项目12',
+        laneId: '1',
+        entryId: '1'
+    },
+    {
+        _id: '13',
+        title: '项目13',
+        laneId: '1',
+        entryId: '1'
+    },
+    {
+        _id: '14',
+        title: '项目14',
+        laneId: '1',
+        entryId: '1'
+    },
+    {
+        _id: '15',
+        title: '项目15',
+        laneId: '2',
+        entryId: '2'
+    },
+    {
+        _id: '16',
+        title: '项目16',
+        laneId: '1',
+        entryId: '1'
+    },
+    {
+        _id: '17',
+        title: '项目17',
+        laneId: '4',
+        entryId: '1'
+    },
+    {
+        _id: '18',
+        title: '项目18',
+        laneId: '',
+        entryId: '1'
+    },
+    {
+        _id: '19',
+        title: '项目19',
+        laneId: '1',
+        entryId: '2'
+    }
+];
+
+export const lanes: ThyBoardLane[] = [
+    {
+        _id: '1',
+        name: '泳道1'
+    },
+    {
+        _id: '2',
+        name: '泳道2'
+    },
+    {
+        _id: '3',
+        name: '泳道3'
+    }
+];