+++ packages/components/board/examples/virtual-scroll/virtual-scroll.component.ts
@@ -1,0 +1,26 @@
+import { Component, OnInit } from '@angular/core';
+import { ThyBoardCard, ThyBoardEntry, ThyBoardLane } from '@tethys/pro/board';
+import { entries, items, lanes } from '../mock';
+
+interface CardInfo extends ThyBoardCard {
+    title: string;
+}
+
+@Component({
+    selector: 'thy-pro-board-virtual-scroll-example',
+    templateUrl: './virtual-scroll.component.html',
+    styleUrls: ['./virtual-scroll.component.scss']
+})
+export class ThyProBoardVirtualScrollExampleComponent implements OnInit {
+    entries: ThyBoardEntry[] = [...entries];
+
+    items: CardInfo[] = [...(items as CardInfo[])];
+
+    lanes: ThyBoardLane[] = [...lanes];
+
+    thyShowLane = true;
+
+    constructor() {}
+
+    ngOnInit(): void {}
+}