+++ packages/components/media/examples/mock.ts
@@ -1,5 +1,6 @@
 export const audioSrc = 'https://dl.espressif.com/dl/audio/ff-16b-2c-44100hz.aac';
 
-export const videoSrc = 'http://clips.vorwaerts-gmbh.de/big_buck_bunny.mp4';
+export const videoSrc =
+    'https://atlas-test.alpha.pingcode.live/file/origin-url?token=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOiI0NTEzNmIwNDIxMzM0ZjYwOGM5ZWVhZDMyZGYzNTkyZSIsInRlYW1faWQiOiI2MjU0ZTY5YjFjOTU3ZGUxODA0ZmRjMTEiLCJwZXJtaXNzaW9uIjoiMTExMTEiLCJmaWxlX2lkIjoiNjY0NWFkMTliYjU2NTA4OTUwZDc0YTAyIiwiaWF0IjoxNzIyODUzNDk0LCJleHAiOjE3MjI5Mzk4OTR9.LROGsniDO7Cum6obx1kP9MTngFpUOcjVO6rWq5xtNgE';
 
 export const videoSrc1 = 'https://dl.espressif.com/dl/audio/ff-16b-2c-44100hz.mp4';